create railroad.rb for text interface, main.rb - file for run app. Delete wagon.rb how as empty file
