+++ App/train_cargo.rb
@@ -1,10 +1,3 @@
 # Класс для создания грузовых поездов
 # Имеет родительский класс Train
-class TrainCargo < Train
-  attr_reader :type, :route
-
-  def initialize(number)
-    super
-    @type = :cargo
-  end
-end
+class TrainCargo < Train; end
