+++ App/train_pass.rb
@@ -1,10 +1,3 @@
 # Класс для создания пассажирских поездов
 # Имеет родительский класс Train
-class TrainPass < Train
-  attr_reader :type, :route
-
-  def initialize(number)
-    super
-    @type = :pass
-  end
-end
+class TrainPass < Train; end
