+++ App/wagon_cargo.rb
@@ -1,9 +1,2 @@
 # Класс для создания вагонов исключительно грузового типа
-class WagonCargo < Wagon
-  attr_reader :type
-
-  def initialize
-    super
-    @type = :cargo
-  end
-end
+class WagonCargo end
