+++ App/wagon_pass.rb
@@ -1,9 +1,2 @@
 # Класс для создания вагонов исключительно пассажирского типа
-class WagonPass < Wagon
-  attr_reader :type
-
-  def initialize
-    super
-    @type = :pass
-  end
-end
+class WagonPass end
